refactor: group src/browser helpers into dom/ and scheduling/ folders
Mirrors the feature-folder layout used in browser-extensions/chrome/src.
pangu.ts and pangu.umd.ts stay at the root since build:types mirrors
src into dist and package.json/cdnjs pin dist/browser/pangu.d.ts and
dist/browser/pangu.umd.js.
