+++ src/browser/dom/visibility-detector.ts
